Update: Move elide facts to client
